Code and bug fix to explicitly control the number of processes python uses.  See config.py in the base BEAST directory to set the number of threads.
Former-commit-id: 0cc85e93d4b31ad97d313f85570b42a574505103

+++ projects/phat_toothpick/fit.py
@@ -576,9 +576,9 @@ def Q_all(lnpfile, sedgrid, qnames, p=[16., 50., 84.], objlist=None, gridbackend
             r['{0:s}_p{1:d}'.format(qname, int(pval))] = per_vals[:,k,i]
 
     r['chi2min'] = chi2_vals
-    r['chi2min_indx'] = chi2_indx
+    r['chi2min_indx'] = chi2_indx.astype(int)
     r['Pmax'] = lnp_vals
-    r['Pmax_indx'] = lnp_indx
+    r['Pmax_indx'] = lnp_indx.astype(int)
 
     if not isinstance(lnpfile, tables.file.File):
         f.close()
